requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.5
httpx[socks]==0.28.1

# Authentication & Security
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.20
pyotp==2.9.0
python-dotenv==1.1.1
pydantic-settings==2.12.0

# Terminal & WebSockets
websockets==15.0.1
ptyprocess==0.7.0
aiofiles==24.1.0

# AI Integration
anthropic==0.69.0

# Utilities
qrcode[pil]==8.2
